refactor(PAT-92): refactor componentes and order comments by time_stamp.
