+++ com/frontend/app/src/components/consultation/display/ConsutationDisplay.jsx
@@ -1,21 +1,10 @@
 import React from 'react';
-import {useEffect, useState} from 'react';
-import { Dialog, DialogTitle, DialogContent, Button, BottomNavigation, BottomNavigationAction, List, ListItem, ListItemText, TextField, Card, CardContent, Typography, Avatar, Box, Grid, IconButton} from '@mui/material';
-import TableContainer from '@mui/material/TableContainer';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableRow from '@mui/material/TableRow';
-import TableCell from '@mui/material/TableCell';
-import {getConsultation, updateConsultationField} from '../../../utils/caseTaker.jsx';
-import ClientTableRow from './ClientTableRow.jsx';
-import EditableFieldRow from '../../EditableFieldRow.jsx';
-import EditableChoiceRow from '../../EditableChoiceRow.jsx';
-import { formatTimestamp } from '../../../utils/tools.jsx';
+import {useState} from 'react';
+import {Dialog, DialogTitle, DialogContent, Button, BottomNavigation, BottomNavigationAction} from '@mui/material';
 import InfoIcon from '@mui/icons-material/Info';
 import CommentIcon from '@mui/icons-material/Comment';
-import AddCircleIcon from '@mui/icons-material/AddCircle';
-import { createComment, getCommentListByConsult } from '../../../utils/comments.jsx';
-import { useUserContext } from '../../../context/UserContext.jsx';
+import Comment from './CommentView.jsx';
+import ConsutationInfoView from './ConsultationInfoView.jsx';
 
 
 /**
@@ -28,148 +17,7 @@ import { useUserContext } from '../../../context/UserContext.jsx';
  * @returns {JSX.Element} - The ConsultationDisplay component JSX.
  */
 const ConsutationDisplay = ({consultation, open, onClose, updateViewTag }) => {
-    const userContext =  useUserContext();
-    const [userData, setUserData] = useState(0);
-    const [consultationData, setConsultation] = useState(consultation)
-    const [updateViewCounter, setUpdateViewCounter] = useState(0); // Force update View
     const [windowNumber, setWindowNumber] = useState(0)
-    const [isFieldsEditing, setIsFieldsEditing] = useState({
-        "description": false,
-        "tag": false,
-        "opponent": false,
-        "progress_state": false,
-    });
-    const [editedFields, setEditedFields] = useState({
-        "description": consultationData.description,
-        "tag": consultationData.tag,
-        "opponent": consultationData.opponent,
-        "progress_state": consultationData.progress_state,
-    });
-    const [fieldsError, setFieldsError] = useState({
-        "description": "",
-        tag: "",
-        "opponent": "",
-        "progress_state": "",
-    });
-    const [comments, setComments] = useState([]);
-    const [newComment, setNewComment] = useState('');
-
-    useEffect(() => {
-        const initComment = async() => {
-            const coments = await getCommentListByConsult(consultation.consultation);
-            setComments(coments);
-        };
-        initComment();
-    }, [consultation.consultation]);
-
-    useEffect(() => {
-        setUserData(userContext.user);
-        console.log(`User ${userContext.user.pk}`);
-    }, [userContext])
-
-    const handleAddComment = async () => {
-        if (newComment.trim() !== '') {
-            const commentData = {user: userData.pk, consultation: consultation.consultation, text: newComment};
-            const commentDict = await createComment(commentData);
-            commentDict.user = userData
-            setComments([commentDict, ...comments]);
-            setNewComment('');
-        }
-    };
-
-    /**
-     * Handles the click event to enable editing of a specific field.
-     * @param {string} fieldKey - The key of the field to be edited.
-     */
-    const handleEditClick = (fieldKey) => {
-        isFieldsEditing[fieldKey] = true;
-        setIsFieldsEditing(isFieldsEditing);
-        fieldsError[fieldKey] = "";
-        setFieldsError(fieldsError)
-        setUpdateViewCounter(updateViewCounter + 1);
-        editedFields[fieldKey] = consultationData[fieldKey]
-        setEditedFields(editedFields)
-    };
-
-    /**
-     * Handles the click event to save changes made to a specific field.
-     * @param {string} fieldKey - The key of the field being edited.
-     */
-    const handleSaveClick = async(fieldKey) => {
-        if (editedFields[fieldKey] === ""
-        || editedFields[fieldKey] === undefined
-        || editedFields[fieldKey] === null
-        ) {
-            fieldsError[fieldKey] = "This field cannot be empty."
-            setFieldsError(fieldsError);
-        } else {
-            try {
-                await updateConsultationField(consultationData.id, fieldKey, editedFields[fieldKey])
-                fieldsError[fieldKey] = "";
-                setFieldsError(fieldsError);
-                setConsultation((prevConsultation) => ({
-                    ...prevConsultation,
-                    [fieldKey]: editedFields[fieldKey],
-                }));
-                isFieldsEditing[fieldKey] = false;
-                setIsFieldsEditing(isFieldsEditing);
-                if (fieldKey === "tag") {
-                    updateViewTag(editedFields.tag)
-                }
-            } catch(e) {
-                fieldsError[fieldKey] = "Error updating field. Please try again."
-                setFieldsError(fieldsError);
-            }
-        }
-        setUpdateViewCounter(updateViewCounter + 1);
-    };
-
-    /**
-     * Handles the change event for a field's value.
-     * @param {Object} event - The change event object.
-     * @param {string} fieldKey - The key of the field being edited.
-     */
-    const handleOnChange = (event, fieldKey) => {
-        const newValue = event.target.value;
-        editedFields[fieldKey] = newValue
-        setEditedFields(editedFields);
-        fieldsError[fieldKey] = "";
-        setFieldsError(fieldsError);
-        setUpdateViewCounter(updateViewCounter + 1);
-    };
-
-    /**
-     * Handles the click event to cancel editing of a specific field.
-     * @param {string} fieldKey - The key of the field to be canceled.
-     */
-    const handleOnCancel = (fieldKey) => {
-        isFieldsEditing[fieldKey] = false;
-        setIsFieldsEditing(isFieldsEditing);
-        fieldsError[fieldKey] = "";
-        setFieldsError(fieldsError);
-        setUpdateViewCounter(updateViewCounter + 1);
-    };
-
-
-    /**
-     * Fetch consultation data if client data is not available.
-     */
-	useEffect(() => {
-		const fetchConsultancy = async () => {
-            try {
-                if (!consultation.client) {
-                    const consultationResponse = await getConsultation(consultation.consultation)
-                    setConsultation(consultationResponse)
-                }
-            } catch (error) {
-                console.error("Failed to fetch Consultation in Card.");
-                console.debug(error);
-            }
-        };
-
-        fetchConsultancy();
-
-    }, [consultation]);
 
     /**
      * Close the dialog.
@@ -204,121 +52,8 @@ const ConsutationDisplay = ({consultation, open, onClose, updateViewTag }) => {
 
 
         <DialogContent>
-        {(windowNumber===0)&&
-        <div elevation={3} style={{ padding: '20px' }}>
-        <TableContainer style={{width: '800px'}}>
-            <Table>
-            <TableBody>
-                <EditableFieldRow
-                    tittle={"Tag:"}
-                    isEditing={isFieldsEditing.tag}
-                    value={isFieldsEditing.tag ? editedFields.tag : consultationData.tag }
-                    onEdit={handleEditClick}
-                    onSave={handleSaveClick}
-                    onChange={handleOnChange}
-                    onCancel={handleOnCancel}
-                    error={fieldsError.tag}
-                    fieldKey={"tag"}
-                />
-                <ClientTableRow clientID={consultationData.client}/>
-                <EditableFieldRow
-                    tittle={"Opponent:"}
-                    isEditing={isFieldsEditing.opponent}
-                    value={isFieldsEditing.opponent ? editedFields.opponent : consultationData.opponent}
-                    onEdit={handleEditClick}
-                    onSave={handleSaveClick}
-                    onChange={handleOnChange}
-                    onCancel={handleOnCancel}
-                    error={fieldsError.opponent}
-                    fieldKey={"opponent"}
-                />
-                <TableRow>
-                <TableCell>Availability State:</TableCell>
-                <TableCell>{consultationData.availability_state}</TableCell>
-                </TableRow>
-                <EditableChoiceRow
-                    title={"Progress State:"}
-                    isEditing={isFieldsEditing.progress_state}
-                    value={consultationData.progress_state}
-                    onEdit={handleEditClick}
-                    onSave={handleSaveClick}
-                    onChange={handleOnChange}
-                    onCancel={handleOnCancel}
-                    error={fieldsError.progress_state}
-                    fieldKey={"progress_state"}
-                    options={["TODO", "IN_PROGRESS", "DONE", "PAUSED", "BLOCKED"]}
-                />
-                <EditableFieldRow
-                    tittle={"Description:"}
-                    isEditing={isFieldsEditing.description}
-                    value={isFieldsEditing.description ? editedFields.description : consultationData.description}
-                    onEdit={handleEditClick}
-                    onSave={handleSaveClick}
-                    onChange={handleOnChange}
-                    onCancel={handleOnCancel}
-                    error={fieldsError.description}
-                    fieldKey={"description"}
-                />
-                <TableRow>
-                <TableCell>Creation time stamp:</TableCell>
-                <TableCell>{formatTimestamp(consultationData.time_stamp)}</TableCell>
-                </TableRow>
-            </TableBody>
-            </Table>
-        </TableContainer>
-        </div>
-        }
-        {(windowNumber===1)&&
-            <div>
-            <Box width="100%" style={{ marginBottom: '10px', display: 'flex', justifyContent: 'flex-end' }}>
-                <TextField
-                    id="outlined-textarea"
-                    placeholder="Placeholder"
-                    label="Add a comment"
-                    multiline
-                    variant="outlined"
-                    fullWidth
-                    value={newComment}
-                    onChange={(e) => setNewComment(e.target.value)}
-                />
-                <IconButton
-                    color="primary"
-                    onClick={handleAddComment}
-                    style={{ marginLeft: '10px' }}
-                    
-                >
-                    <AddCircleIcon />
-                </IconButton>
-            </Box>
-            {comments.map((comment, index) => (
-                <Box key={comment.id} width="100%" style={{ marginBottom: '10px' }}>
-                <Card key={comment.id} variant="outlined" style={{ marginBottom: '10px' }}>
-                <CardContent sx={{ whiteSpace: 'pre-line' }}>
-                <Grid container spacing={2}>
-                    <Grid item>
-                        <Avatar sx={{ bgcolor: 'primary.main' }}>
-                            {comment.user.username?.charAt(0)}
-                        </Avatar>
-                        </Grid>
-                        <Grid item>
-                        <Typography variant="h6" component="div">
-                            {comment.user.username}
-                        </Typography>
-                    </Grid>
-                    <Grid item>
-                        <Typography>
-                            {comment.text}
-                        </Typography>
-                    </Grid>
-                </Grid>
-            </CardContent>
-                </Card>
-                </Box>
-            ))}
-
-            </div>
-        }
-
+        {(windowNumber===0)&&<ConsutationInfoView consultation={consultation} updateViewTag={updateViewTag}/>}
+        {(windowNumber===1)&& <Comment consultationID={consultation.consultation}/>}
         </DialogContent>
         <Button onClick={handleClose} color="primary">
             Close
